Move build_html_tree_highlighting_codes to presenters.py
