+++ codelists/tests/factories.py
@@ -24,15 +24,20 @@ class SignOffFactory(factory.django.DjangoModelFactory):
     date = factory.fuzzy.FuzzyDate(datetime.date(2020, 1, 1))
 
 
-def create_codelist():
+def create_codelist(csv_data=None):
     p = ProjectFactory()
+    if csv_data is None:
+        csv_data = (
+            "code,description\n1067731000000107,Injury whilst swimming (disorder)"
+        )
+
     return actions.create_codelist(
         project=p,
         name="Test Codelist",
         coding_system_id="snomedct",
         description="This is a test",
         methodology="This is how we did it",
-        csv_data="code,description\n1067731000000107,Injury whilst swimming (disorder)",
+        csv_data=csv_data,
     )
