+++ codelists/views.py
@@ -24,6 +24,7 @@ from .forms import (
 )
 from .hierarchy import Hierarchy
 from .models import Codelist, CodelistVersion
+from .presenters import build_html_tree_highlighting_codes
 
 
 def index(request):
@@ -277,7 +278,7 @@ def version(request, project_slug, codelist_slug, qualified_version_str):
         html_definition = build_html_definition(coding_system, hierarchy, definition)
         if clv.coding_system_id in ["ctv3", "ctv3tpp"]:
             subtree = tree_utils.build_subtree(coding_system, clv.codes)
-            html_tree = tree_utils.build_html_tree_highlighting_codes(
+            html_tree = build_html_tree_highlighting_codes(
                 coding_system, hierarchy, subtree, definition
             )
         else:
